requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.4
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.2.0
    # via anyio
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.16
    # via -r requirements.in
langchain-community==0.0.32
    # via
    #   -r requirements.in
    #   langchain
langchain-core==0.1.42
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.3
    # via -r requirements.in
langchain-text-splitters==0.0.1
    # via langchain
langsmith==0.1.47
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.21.1
    # via dataclasses-json
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.17.1
    # via langchain-openai
orjson==3.10.0
    # via langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pydantic==2.7.0
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.18.1
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   tiktoken
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.29
    # via
    #   langchain
    #   langchain-community
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.6.0
    # via langchain-openai
tqdm==4.66.2
    # via openai
typing-extensions==4.11.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.1
    # via requests
yarl==1.9.4
    # via aiohttp
